feat(contrib): add a basic evaluator script for tests
Evaluates Nixpkgs in 1627 on my Ryzen Zen 4 for a result of 652MB of jsonl (41.8MB zstd compressed).
We are forced to recurse because recurseForDerivations are removed in this instance.

Ideally, we do like the rebuild-amount.sh of nixpkgs in maintainers/ and we tweak the
attribute set to remove parts of it and keep only what matters.

Anyway, that's good enough as a start, pick your copy here:
https://files.lahfa.xyz/private/evaluation.jsonl.zstd