feat(command): manual ingestion of an evaluation result
You can see, after many tinkering of the real data, how we came to
make a lot of models less "strict" in terms of validation.
We should move the question to a Tracking Issue inside nixpkgs on the data integrity
of e.g. licenses or meta attributes.
Either case, this enables ingesting an evaluation result in a reasonable amount of time, it _could_ be
faster but at the cost of a much more complicated bulk creation code.
With this, we can evaluate, ingest this manually in a cron and move on.