feat: keep CVE data in a local cache directory for efficient reimports (#92)
Most of the time spent on CVE import is fetching and uncompressing the
CVE bundle from Github. This penalizes development work on the database,
where multiple reimports might be needed.

Now the CVE data gets stored in a local cache directory after the first fetch.
It's still possible to force the deletion of the CVE cache to download the
bundle again with the `--force-download` option.