feat: make the realtime evaluation faster
The strategy here is to have a performance that is bounded by the
execution time of `nix-eval-jobs` + some constant overhead due to finishing the last batch
insertion.

If we say that for example, it takes 400s to consume an average batch, in those 400s,
you may get a batch of comparable size, you will indeed do execution time of `nej` + 400s
for realtime ingestion.

This is a reasonable result for realtime ingestion, note that we could attempt to perform
bulk insertion to reduce that constant down, but at the cost of a more complicated piece of code.

For that, we steal available results and bulk load them as fast as we
can. The "slow" path is the maintainer and licenses ingestion and check
because it's hard to bulk load it further.

On my AMD Ryzen 9 7900X 12-Core Processor with fast NVMe, I can
*evaluate* and *ingest* a channel in 807 (~ 15 minutes) seconds.