feat: nixpkgsissue activity log view
Improve the messy representation of the different `pghistory` event tables:

- Create a view `log_nixpkgsissue` in the PostgreSQL backend and introduce an unmanaged model
`NixpkgsIssueLogView` to display a list of changes per timestamp and user.
- Register `NixpkgsIssueLogView` in the `pghistory` app to declutter the log reporting from the
already crowded `shared` section.
- Delete the previous activitiy log models and views in favour of the unified view model.

If performance starts being problematic for the `log_nixpkgsissue` view, we can
make it a materialized view and add a trigger to update its contents on `shared.nixpkgsissue`
and it's m2m relationships updates.