feat: introduce provenance for derivation candidates
In the future, we will be able to bolt on more fields to take notes
which embedding has been used, which search term has been used, etc,
etc.

For now, we have a provenance flags we can use to hint the frontend what
has been used.

The dance for the M2M is required because it's not trivial to migrate
everything.

Progress towards:

- https://github.com/Nix-Security-WG/nix-security-tracker/issues/189
- https://github.com/Nix-Security-WG/nix-security-tracker/issues/274

Signed-off-by: Raito Bezarius <masterancpp@gmail.com>