ci: pin Nixpkgs for deployments
this doesn't rely on NIX_PATH and is therefore fully pure, and uses our
own, controlled dependencies.

the Nixpkgs for deployment currently not decoupled from the one for
development because the build is broken on the newest version.

we still need a policy for how to run updates, but ideally we'd use
exactly one version of Nixpkgs for everything.