feat: perform status changes on suggestions via AJAX
This also allows for an easy undo workflow, as the suggestion collapses
into a one-line tile that has an undo button.

This aims to still maintain a non-JS workflow that works as just before.

For achieving this, the template for a single suggestion was refactored
out of the suggestion list.