Define publish as a status change
Publishing an issue was developed as a separate workflow, for no good
reason. Now that we have a new "published" status for suggestions, we
can just see publication as the ultimate status change, and take
advantage of the existing infra (such as the component showing a status
change to the user).

Additionally, we extend this status change component to open a new tab
with the GitHub issue on the side.