fix: add toggle to disable GitHub sync
Make explicit that this needs to be set deliberately:
Enabling or disabling it by default will break either tests or
production at runtime if the caller doesn't know the setting exists.
Making it required will at least inform the caller early on that there is a decision to be made.

This fixes the VM tests (and showcases how cool `pydantic-settings` is).