feat: re-enable "Publish issue" button (#523)
* feat: published status in the view

Add a "Publish issue" button to update a suggestion from accepted to
published. Update the status change component to link to the opened
Github issue (instead of the undo button). Note that this commit doesn't
actually implement any backend logic, so the button has no effect
currently.

* feat: handle issue publication

This commit finally ties the knot, using previously introduced endpoints
and helper to implement the "Publish issue" action. When clicked, the
status of the suggestion is changed to "published", a NixpkgsIssue
object is created, and finally a Github issue is created on the Nixpkgs
repository.